move all test utils to smartContractUtils
